# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Beauty Salon CRM API",
        "description": "API для системи управління салоном краси",
        "contact": {
            "email": "info@beautysalon.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/appointments-calendar": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Отримати записи для календаря",
                "description": "Повертає записи у форматі FullCalendar для відображення в календарі",
                "operationId": "3c20be6fead78f3b4536319c66a42ae3",
                "parameters": [
                    {
                        "name": "employee_id",
                        "in": "query",
                        "description": "ID майстра для фільтрації",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Статус запису",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "scheduled",
                                "confirmed",
                                "completed",
                                "cancelled"
                            ],
                            "example": "scheduled"
                        }
                    },
                    {
                        "name": "start",
                        "in": "query",
                        "description": "Початкова дата діапазону",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-12-01"
                        }
                    },
                    {
                        "name": "end",
                        "in": "query",
                        "description": "Кінцева дата діапазону",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-12-31"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успішна відповідь",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Іван Іванов - Стрижка"
                                            },
                                            "start": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-12-25T10:00:00+00:00"
                                            },
                                            "end": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-12-25T11:00:00+00:00"
                                            },
                                            "color": {
                                                "type": "string",
                                                "example": "#667eea"
                                            },
                                            "extendedProps": {
                                                "properties": {
                                                    "client_name": {
                                                        "type": "string",
                                                        "example": "Іван Іванов"
                                                    },
                                                    "employee_name": {
                                                        "type": "string",
                                                        "example": "Майстер"
                                                    },
                                                    "service_name": {
                                                        "type": "string",
                                                        "example": "Стрижка"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "scheduled"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 800
                                                    },
                                                    "notes": {
                                                        "type": "string",
                                                        "example": "Примітки",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизовано"
                    },
                    "403": {
                        "description": "Доступ заборонено"
                    }
                },
                "security": [
                    {
                        "session": []
                    }
                ]
            }
        },
...
        "/api/services/{service}/employees": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Отримати майстрів за послугою",
                "description": "Повертає список активних майстрів, які надають вказану послугу",
                "operationId": "fe01532424332e7eba8fc6c1d96bb256",
                "parameters": [
                    {
                        "name": "service",
                        "in": "path",
                        "description": "ID послуги",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успішна відповідь",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Майстер"
                                            },
                                            "specialization": {
                                                "type": "string",
                                                "example": "Стрижка",
                                                "nullable": true
                                            },
                                            "rating": {
                                                "type": "string",
                                                "example": "4.8"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Послуга не знайдена"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "session": {
                "type": "apiKey",
                "description": "CSRF Token для автентифікації",
                "name": "X-CSRF-TOKEN",
                "in": "header"
            }
        }
    },
...
}